requirements.txt
opencv-python
numpy
onnx
onnxruntime
onnxruntime-gpu
flask
Pillow
